requirements.txt
fastapi==0.116.1
httpx==0.28.1  # critical
python-dotenv==1.1.1  # critical
beautifulsoup4==4.13.4
python-docx==1.2.0
fpdf==1.7.2
pytest-asyncio==1.1.0
pytest-cov==6.2.1
ruff==0.12.7
black==25.1.0
pydocstyle==6.3.0
pdoc==15.0.4
